move types to linked_time_fob/
