+++ tensorboard/webapp/metrics/BUILD
@@ -47,7 +47,7 @@ tf_ts_library(
     visibility = ["//tensorboard/webapp/metrics:__subpackages__"],
     deps = [
         "//tensorboard/webapp/widgets/histogram:types",
-        "//tensorboard/webapp/widgets/linked_time_types",
+        "//tensorboard/webapp/widgets/linked_time_fob:types",
     ],
 )
