+++ tensorboard/webapp/widgets/histogram/histogram_component.ts
@@ -30,7 +30,7 @@ import {takeUntil} from 'rxjs/operators';
 import * as d3 from '../../third_party/d3';
 import {HCLColor} from '../../third_party/d3';
 import {AxisDirection} from '../linked_time_fob/linked_time_fob_controller_component';
-import {LinkedTime} from '../linked_time_types/types';
+import {LinkedTime} from '../linked_time_fob/linked_time_types';
 import {
   Bin,
   HistogramData,
