+++ tensorboard/webapp/widgets/linked_time_fob/BUILD
@@ -28,8 +28,8 @@ tf_ng_module(
         ":linked_time_fob_controller_styles",
     ],
     deps = [
+        ":types",
         "//tensorboard/webapp/third_party:d3",
-        "//tensorboard/webapp/widgets/linked_time_types",
         "@npm//@angular/common",
         "@npm//@angular/core",
     ],
@@ -44,12 +44,19 @@ tf_ts_library(
     ],
     deps = [
         ":linked_time_fob",
+        ":types",
         "//tensorboard/webapp/angular:expect_angular_core_testing",
         "//tensorboard/webapp/testing:dom",
         "//tensorboard/webapp/third_party:d3",
-        "//tensorboard/webapp/widgets/linked_time_types",
         "@npm//@angular/core",
         "@npm//@angular/platform-browser",
         "@npm//@types/jasmine",
+    ],
+)
+
+tf_ts_library(
+    name = "types",
+    srcs = [
+        "linked_time_types.ts",
     ],
 )
