+++ tensorboard/webapp/widgets/linked_time_fob/linked_time_fob_controller_component.ts
@@ -23,7 +23,7 @@ import {
   ViewChild,
 } from '@angular/core';
 import {ScaleLinear, ScaleTime} from '../../third_party/d3';
-import {LinkedTime} from '../linked_time_types/types';
+import {LinkedTime} from './linked_time_types';
 
 export enum AxisDirection {
   HORIZONTAL,
