--- tensorboard/webapp/widgets/linked_time_types/BUILD
@@ -1,10 +1,0 @@
-load("//tensorboard/defs:defs.bzl", "tf_ts_library")
-
-package(default_visibility = ["//tensorboard:internal"])
-
-tf_ts_library(
-    name = "linked_time_types",
-    srcs = [
-        "types.ts",
-    ],
-)
